# Don't log in bigquery results hotpath